Add configuration option to skip splitting replies
For emails with large bodies, splitting the replies out can take an
excessive amount of time. This new configuration option gives the app
developer control over whether or not the replies are split out of the
email body, the default behavior it to always split replies.

To skip splitting replies, add this to the app initializer:

Griddler.configure do |config|
  config.split_replies = false
end